@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "squadlanes-extraction"
-version = "2.3.0"
+version = "2.3.1"
 description = "Datamining scripts used to extract map images and lane information from Squad game files"
 authors = ["Tim Schmidt <tim.schmidt@khorne.faith>"]
 license = "AGPLv3"
@@ -11,7 +11,7 @@ extract = "squadlanes_extraction.main:extract"
 tiles = "squadlanes_extraction.main:tiles"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 PyYAML = "^5.4.1"
 pwntools = "^4.5.0"
 
